# openapi/web-v1.json
{
  "components": {
    "schemas": {
...
      "IdentityResponse": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_admin": {
            "title": "Is Admin",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email",
          "is_admin"
        ],
        "title": "IdentityResponse",
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "password": {
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "recallum_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "Recallum Web API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/login": {
      "post": {
        "operationId": "login_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Validation error"
          },
          "429": {
            "description": "Too many requests"
          }
        },
        "security": [],
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "logout_auth_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Validation error"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Logout",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "me_auth_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "413": {
            "description": "Request body too large"
          },
          "422": {
            "description": "Validation error"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Me",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}